Fix configuration to run in Elastic Beanstalk
It looks like Elastic Beanstalk will not run the contets of the
current environment, but instead the last commit. I checked the
ECS by SSHing and top'ing and saw that the Nginx process was not
running and that npm run test|start were running. So the development
Docker environment were running in the EC2.

This commit updates the configuration and commits so that EB will
use this new commit.